Basic custom Ogg Opus comment tag write support
These changes let PyOgg users add custom comment tags to the generated
stream, as opposed to the current behavior of hardcoding no comment tags
on all cases, which can come in handy for stream identification
requirements.

While at it, I've tweaked the vendor string from `ENCODER=PyOgg` to
`PyOgg <VERSION>`, where `<VERSION>` is the value of the `__version__`
variable at `__init__.py`. The rationale for this change is that this
brings the vendor string format more in line with the conventions
used by libopus and other encoders (such as FFmpeg and libavcodec),
while also being more useful to track the precise encoder code used to
generate the stream thanks to the addition of the version number.